debug_grounding_dino/requirements.txt
einops
pandas==1.5.2
ptflops
matplotlib
psutil==5.6.7
tqdm==4.48.2
scikit_learn==0.24.1
torchscan==0.1.1
googledrivedownloader==0.4
requests==2.26.0
timm==0.9.12
transformers==4.8.2
faiss_gpu
Pillow==10.1.0
folium
plotly
albumentations
opencv-python-headless==4.8.0.74
opencv-python==4.8.0.74
kornia
h5py
fast_pytorch_kmeans
numpy==1.22.2
xgboost==1.6.2
mlflow==2.3.1
hydra-core==1.3.2
hydra-joblib-launcher==1.2.0
addict
yapf
supervision